add support for --check/--diff to:
- hashivault_auth_ldap.py
- hashivault_auth_method.py
- hashivault_k8s_auth_config.py
- hashivault_ldap_group.py
- hashivault_policy.py
- hashivault_secret_engine.py
- hashivault_userpass.py

additional notes:

- hashivault_userpass.py now uses new hvac client.auth.userpass methods,
instead of methods directly on the client object that will be deprecated
in hvac 1.0.0.

- hashivault_auth_ldap.py remove spurious parameters token_ttl and
token_max_ttl, which are not returned by the API (or apparently consumed
by them either), so don't include them in current/desired state or
args processing.

add new modules:

- hashivault_aws_secret_engine_{config,role}.py
- hashivault_ldap_{group,user}_list.py
- hashivault_ldap_user.py
- hashivault_userpass_list.py

additional notes:

- hashivault_aws_secret_engine_role.py supports legacy_params mode in
which hvac/Vault only pays attention to policy_document/policy_arns and
other parameters are ignored; on write the legacy format is converted
into the current role structure, in so doing setting credential_types to
[iam_user, federation_token].